language: create files for model and token
