+++ libs/language/src/lib/translation-loader/index.ts
@@ -1,1 +1,3 @@
+export * from './token';
 export * from './translation-loader';
+export * from './translations.model';
