+++ libs/language/src/lib/translation-loader/token.ts
@@ -1,0 +1,7 @@
+import { InjectionToken } from '@angular/core';
+
+import { Translations } from './translations.model';
+
+export const TRANSLATIONS_TOKEN = new InjectionToken<Translations>(
+  'Object containing translations'
+);
