+++ libs/language/src/lib/translation-loader/translation-loader.ts
@@ -1,15 +1,9 @@
-import { Inject, InjectionToken } from '@angular/core';
+import { Inject } from '@angular/core';
 import { TranslateLoader } from '@ngx-translate/core';
 import { from, Observable } from 'rxjs';
 
-export interface Translations {
-  'en-US': any;
-  'pt-BR': any;
-}
-
-export const TRANSLATIONS_TOKEN = new InjectionToken<Translations>(
-  'Object containing translations'
-);
+import { TRANSLATIONS_TOKEN } from './token';
+import { Translations } from './translations.model';
 
 export class LibTranslationLoader implements TranslateLoader {
   public constructor(
