+++ libs/language/src/lib/translation-loader/translations.model.ts
@@ -1,0 +1,4 @@
+export interface Translations {
+  'en-US': any;
+  'pt-BR': any;
+}
